requirements.txt
acres==0.2.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
async-timeout==4.0.3
asyncer==0.0.7
attrs==24.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
chainlit==1.3.2
charset-normalizer==3.4.1
chevron==0.14.0
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.31.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-google-genai==2.0.7
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.10
Lazify==0.4.0
literalai==0.0.623
llama-cloud==0.1.7
llama-index==0.12.9
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.10
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.12
llama-index-multi-modal-llms-openai==0.4.2
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.2
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.19
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.20.1
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numpy==1.26.4
ollama==0.4.5
openai==0.28.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.13
packaging==23.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
prov==2.0.1
psutil==6.1.1
pure_eval==0.2.3
puremagic==1.28
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.25.1
pyparsing==3.2.1
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.9
python-socketio==5.12.1
pytz==2024.2
pywin32==308
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
simple-websocket==1.1.0
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
striprtf==0.0.26
sympy==1.13.1
syncer==2.0.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.28.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.25.0
watchdog==6.0.0
watchfiles==0.20.0
wcwidth==0.2.13
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
